* introduction of a basic gui system * gui facilities * FIXME: fix incorrect render target rendering when the UIPass is not plugged in (discovered this was an existing issue which happens whenever the UIPass and predominantly its SpriteFont and SpriteBatch and not bound to the pipeline; it has to be fixed with NO1 priority) * fixed various warnings * R3ctangle class split into 2 classes: RectangleF and RectangleI with floats and ints member data respectively * bugfixes with util functions * deactivated TransparentPass from renderer and cube, TODO: will fix it

+++ KeyEngine/inc/geometry.h
@@ -17,7 +17,7 @@ struct Geometry
 	static TriangleMesh makePlanarGridTextured( const int length = 2, const int width = 2, int nDivisionsX = 1, int nDivisionsY = 1 );
 	//	\function	makePlanarGridTexturedFromHeighmap	||	\date	2022/11/21 0:06
 	//	\brief	engine unit is cm, terrain has to be relatively large without consuming a vast amount of memory with indices and vertices, so we multiply its area by `terrainAreaUnitMultiplier`
-	static TriangleMesh makePlanarGridTexturedFromHeighmap( const std::string &filename, const int normalizeAmount = 4, const int terrainAreaUnitMultiplier = 10, const float length = 2.0f, const float width = 2.0f, int nDivisionsX = 1, int nDivisionsY = 1 );
+	static TriangleMesh makePlanarGridTexturedFromHeighmap( const std::string &filename, const int normalizeAmount = 4, const int terrainAreaUnitMultiplier = 10, const int length = 2, const int width = 2, int nDivisionsX = 1, int nDivisionsY = 1 );
 	static TriangleMesh makeCameraFrustum( const float width, const float height, const float nearZ, const float farZ );
 	static TriangleMesh makeCameraWidget();
 };
